extract window.open to a helper method
