+++ src/app/shared/helpers/navigation/ctrl-click.helpers.ts
@@ -1,7 +1,9 @@
+import { openUrlInNewTab } from './navigation.helpers';
+
 export function verifyClickAndOpenNewTab(event: MouseEvent, url: string): boolean {
   if (!event.ctrlKey && event.button !== 1) return false;
 
-  window.open(url, '_blank');
+  openUrlInNewTab(url);
   event.preventDefault();
   event.stopImmediatePropagation();
   return true;
