+++ src/app/shared/helpers/navigation/navigation.helpers.ts
@@ -1,0 +1,3 @@
+export function openUrlInNewTab(url: string): void {
+  window.open(url, '_blank');
+}
